Add `EventSettings` and organize files
